+++ web/yaamp/core/functions/yaamp.php
@@ -618,7 +618,7 @@ function yaamp_user_rate_bad($userid, $algo = null)
         $interval = yaamp_hashrate_step();
         $delay    = time() - $interval;
 
-        $diff = (double) controller()->memcache->get_database_scalar("yaamp_user_diff_avg-$userid-$algo", "SELECT avg(difficulty) FROM shares WHERE valid AND time>$delay AND userid=$userid AND algo=:algo", array(
+        $diff = (float) controller()->memcache->get_database_scalar("yaamp_user_diff_avg-$userid-$algo", "SELECT avg(difficulty) FROM shares WHERE valid AND time>$delay AND userid=$userid AND algo=:algo", array(
             ':algo' => $algo
         ));
 
@@ -666,7 +666,7 @@ function yaamp_worker_rate_bad($workerid, $algo = null)
         $interval = yaamp_hashrate_step();
         $delay    = time() - $interval;
 
-        $diff = (double) controller()->memcache->get_database_scalar("yaamp_worker_diff_avg-$workerid-$algo", "SELECT avg(difficulty) FROM shares WHERE valid AND time>$delay AND workerid=" . $workerid);
+        $diff = (float) controller()->memcache->get_database_scalar("yaamp_worker_diff_avg-$workerid-$algo", "SELECT avg(difficulty) FROM shares WHERE valid AND time>$delay AND workerid=" . $workerid);
 
         $rate = dboscalar("SELECT count(id) FROM shares WHERE valid!=1 AND time>$delay AND workerid=" . $workerid);
 
@@ -771,7 +771,7 @@ function yaamp_job_rate_bad($jobid)
         $interval = yaamp_hashrate_step();
         $delay    = time() - $interval;
 
-        $diff = (double) controller()->memcache->get_database_scalar("yaamp_job_diff_avg-$jobid", "SELECT avg(difficulty) FROM jobsubmits WHERE valid AND time>$delay AND jobid=" . $jobid);
+        $diff = (float) controller()->memcache->get_database_scalar("yaamp_job_diff_avg-$jobid", "SELECT avg(difficulty) FROM jobsubmits WHERE valid AND time>$delay AND jobid=" . $jobid);
 
         $rate = dboscalar("SELECT count(id) FROM jobsubmits WHERE valid!=1 AND time>$delay AND jobid=" . $jobid);
         $rate = $rate * $diff * $target / $interval / 1000;
