+++ web/yaamp/modules/site/results/graph_market_balance.php
@@ -13,7 +13,7 @@ $markets = dbolist("SELECT M.id, M.name, M.priority, MIN(MH.balance) AS min, MAX
 	GROUP BY M.id, M.name, M.priority HAVING max > 0
 	ORDER BY M.priority DESC, M.name");
 
-$stackedMax = (double)0;
+$stackedMax = (float)0;
 
 $series = array();
 foreach ($markets as $m)
@@ -29,7 +29,7 @@ foreach ($markets as $m)
         $d = date('Y-m-d H:i', $histo->time);
         $series[$m['name']][] = array(
             $d,
-            (double)bitcoinvaluetoa($histo->balance)
+            (float)bitcoinvaluetoa($histo->balance)
         );
 
         $max = max($max, $histo->balance);
@@ -47,7 +47,7 @@ foreach ($stats as $histo)
     $d = date('Y-m-d H:i', $histo->time);
     $series[YAAMP_SITE_NAME][] = array(
         $d,
-        (double)bitcoinvaluetoa($histo->balance)
+        (float)bitcoinvaluetoa($histo->balance)
     );
     $max = max($max, $histo->balance);
 }
@@ -113,6 +113,6 @@ foreach ($series as $name => $serie)
 echo json_encode(array(
     'data' => array_values($series) ,
     'labels' => array_keys($series) ,
-    'rangeMin' => (double)0.0,
+    'rangeMin' => (float)0.0,
     'rangeMax' => ($stackedMax * 1.10) ,
 ));
